updated alertNotification and defaultlayout
