+++ frontend/src/styles/_components.scss
@@ -4,4 +4,5 @@
 @use 'components/child';
 @use 'components/summary';
 @use 'components/dialog'; 
-@use 'components/cards'; 
+@use 'components/cards';
+@use 'components/alerts';
