+++ frontend/src/styles/components/_base.scss
@@ -11,6 +11,50 @@ body, .v-application {
   );
 }
 
+.app-background {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: linear-gradient(135deg, 
+    #faf7f2 0%,
+    #f8f4ef 25%,
+    #fcfaf7 50%,
+    #f5f2ed 75%,
+    #faf7f2 100%
+  );
+  z-index: 0;
+
+  &::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: url("data:image/svg+xml,%3Csvg width='100' height='100' viewBox='0 0 100 100' xmlns='http://www.w3.org/2000/svg'%3E%3Cg fill='%23FFB6C1' fill-opacity='0.02'%3E%3Cpolygon points='50 0 60 40 100 50 60 60 50 100 40 60 0 50 40 40'/%3E%3C/g%3E%3C/svg%3E");
+    opacity: 0.6;
+  }
+}
+
+.custom-app {
+  position: relative;
+  overflow: hidden;
+}
+
+.custom-main {
+  padding-left: 110px ;
+  padding-top: 72px ;
+  position: relative;
+  z-index: 1;
+}
+.custom-main-content {
+  min-height: calc(99vh - 72px);
+  padding: 23px;
+  position: relative;
+}
+
 .left-section {
   display: flex;
   align-items: center;
