+++ frontend/src/styles/components/_dialog.scss
@@ -2,7 +2,7 @@
 
 // Form field styles
 .dialog-field {
-  margin-bottom: 8px;
+  margin-bottom: $spacing-sm;
   
   .v-field__input {
     color: $field-text;
@@ -22,6 +22,15 @@
   
   &.v-field--focused .v-field__outline {
     border-color: $field-border-focus;
+  }
+
+  :deep(.v-field) {
+    background: $field-background;
+    border-radius: $border-radius-lg;
+  }
+  
+  :deep(.v-field--focused) {
+    background: $field-background;
   }
 }
 
@@ -86,23 +95,81 @@
 
 // Dialog Layout
 .dialog-card {
-  background-color: $dialog-background;
-  color: $dialog-text;
+  background: $dialog-background;
+  backdrop-filter: blur(20px);
+  border: 1px solid $dialog-border;
+  border-radius: $border-radius-xl;
+  box-shadow: $shadow-lg;
+  overflow: hidden;
 }
 
 .dialog-title {
-  background-color: $dialog-background;
-  color: $dialog-text;
+  background: linear-gradient(135deg, 
+    rgba($app-primary, 0.1) 0%,
+    rgba($app-primary-light, 0.05) 100%
+  );
   border-bottom: 1px solid $dialog-border;
+  padding: $spacing-lg;
+}
+
+.dialog-title-content {
+  display: flex;
+  align-items: center;
+  gap: $spacing-md;
+  width: 100%;
+}
+
+.dialog-title h3 {
+  font-family: $font-heading;
+  font-size: 20px;
+  font-weight: 600;
+  color: $dialog-text;
+  margin: 0;
+}
+
+.child-chip {
+  font-family: $font-child;
+  font-weight: 500;
 }
 
 .dialog-content {
-  background-color: $dialog-background;
-  color: $dialog-text;
-  padding: 24px;
+  padding: $spacing-lg;
 }
 
 .dialog-actions {
-  padding: 16px 24px 24px 24px;
-  gap: 12px;
+  padding: $spacing-md $spacing-lg $spacing-lg $spacing-lg;
+  gap: $spacing-md;
+}
+
+.cancel-btn {
+  color: rgba($app-text-primary, 0.7);
+  font-weight: 500;
+}
+
+.save-btn {
+  font-weight: 600;
+  border-radius: $border-radius-lg;
+  text-transform: none;
+  letter-spacing: -0.01em;
+}
+
+.alert-content {
+  font-family: $font-primary;
+}
+
+.alert-line {
+  font-size: 14px;
+  line-height: 1.4;
+  margin-bottom: $spacing-xs;
+  
+  &:last-child {
+    margin-bottom: 0;
+  }
+}
+
+// Modern dialog overlay
+.modern-dialog {
+  :deep(.v-overlay__content) {
+    background: transparent;
+  }
 }
