+++ frontend/src/styles/components/_navigation.scss
@@ -104,6 +104,7 @@
     bottom: 0;
     background: url("data:image/svg+xml,%3Csvg width='40' height='40' viewBox='0 0 40 40' xmlns='http://www.w3.org/2000/svg'%3E%3Cg fill='%23ffffff' fill-opacity='0.03'%3E%3Ccircle cx='5' cy='5' r='3'/%3E%3Ccircle cx='35' cy='35' r='3'/%3E%3C/g%3E%3C/svg%3E");
     opacity: 0.4;
+  background: white;
   }
 }
